Transaction: Refactor for transaction package move
