+++ app/src/main/java/ua/safetynet/payment/Transaction.java
@@ -1,11 +1,8 @@
-package ua.safetynet;
+package ua.safetynet.payment;
 
 import com.google.firebase.firestore.IgnoreExtraProperties;
-import com.google.firebase.firestore.ServerTimestamp;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
